+++ src/main/java/br/com/fiap/GreenHouse/model/Tarefa.java
@@ -4,14 +4,15 @@ import java.time.LocalDate;
 
 public class Tarefa {
 
-   private int id;
+   private Long id;
    private String nome;
    private String descricao;
    private int experiencia;
    private int idCategoria; 
    private LocalDate data;
 
-public Tarefa(int id, String nome, String descricao, int experiencia, int idCategoria, LocalDate data) {
+
+public Tarefa(Long id, String nome, String descricao, int experiencia, int idCategoria, LocalDate data) {
     this.id = id;
     this.nome = nome;
     this.descricao = descricao;
@@ -21,14 +22,20 @@ public Tarefa(int id, String nome, String descricao, int experiencia, int idCate
 }
 
 
-public int getId() {
+
+
+public Long getId() {
     return id;
 }
 
 
-public void setId(int id) {
+
+
+public void setId(Long id) {
     this.id = id;
 }
+
+
 
 
 public String getNome() {
@@ -36,9 +43,13 @@ public String getNome() {
 }
 
 
+
+
 public void setNome(String nome) {
     this.nome = nome;
 }
+
+
 
 
 public String getDescricao() {
@@ -46,9 +57,13 @@ public String getDescricao() {
 }
 
 
+
+
 public void setDescricao(String descricao) {
     this.descricao = descricao;
 }
+
+
 
 
 public int getExperiencia() {
@@ -56,9 +71,13 @@ public int getExperiencia() {
 }
 
 
+
+
 public void setExperiencia(int experiencia) {
     this.experiencia = experiencia;
 }
+
+
 
 
 public int getIdCategoria() {
@@ -66,9 +85,13 @@ public int getIdCategoria() {
 }
 
 
+
+
 public void setIdCategoria(int idCategoria) {
     this.idCategoria = idCategoria;
 }
+
+
 
 
 public LocalDate getData() {
@@ -76,9 +99,13 @@ public LocalDate getData() {
 }
 
 
+
+
 public void setData(LocalDate data) {
     this.data = data;
 }
+
+
 
 
 @Override
